Untested -- Implement shadow registers for speed. Assimilate SPI file into TMC5272 driver.
The lag / latency issue comes from redundant SPI writes. To solve this, we need to check whether a SPI write uses the same values as already stored on the device; however, this requires either reading the value in a SPI read transaction (bad) or storing it locally as a shadow register (better). This commit does that.

This change also revealed that the TMC5272_SPI file is unnecessary encapsulation -- it ends up in circular inclusions. Merged these two files into one.